payment method that may or may not work

+++ src/main/java/com/uq/jokievents/model/LocalityOrder.java
@@ -19,5 +19,5 @@ public class LocalityOrder {
 
     private int numTicketsSelected;
     private String localityName;
-    private Double totalPaymentAmount;
+    private float totalPaymentAmount;
 }
